Simplify crypto exchange plugin
