+++ examples/minimal/boot.js
@@ -24,7 +24,7 @@ import 'meteor/unchained:core-pricing/plugins/order-discount';
 import 'meteor/unchained:core-pricing/plugins/order-delivery';
 import 'meteor/unchained:core-pricing/plugins/order-payment';
 import 'meteor/unchained:core-pricing/plugins/product-catalog-price';
-import 'meteor/unchained:core-pricing/plugins/product-price-fiat-crypto-exchange';
+import 'meteor/unchained:core-pricing/plugins/product-price-coinbase-exchange';
 import 'meteor/unchained:core-pricing/plugins/product-discount';
 import 'meteor/unchained:core-pricing/plugins/product-swiss-tax';
 import 'meteor/unchained:core-pricing/plugins/delivery-swiss-tax';
