--- packages/core-pricing/constants/available-exchanges.js
@@ -1,50 +1,0 @@
-const AVAILABLE_CURRENCIES = {
-  fiat: [
-    'USD',
-    'JPY',
-    'BGN',
-    'CZK',
-    'DKK',
-    'GBP',
-    'HUF',
-    'PLN',
-    'RON',
-    'SEK',
-    'CHF',
-    'ISK',
-    'NOK',
-    'HRK',
-    'RUB',
-    'TRY',
-    'AUD',
-    'BRL',
-    'CAD',
-    'CNY',
-    'HKD',
-    'IDR',
-    'ILS',
-    'INR',
-    'KRW',
-    'MXN',
-    'MYR',
-    'NZD',
-    'PHP',
-    'SGD',
-    'THB',
-    'ZAR',
-  ],
-  crypto: [
-    'BTC',
-    'ETH',
-    'XRP',
-    'USDT',
-    'BCH',
-    'BSV',
-    'LTC',
-    'EOS',
-    'BNB',
-    'XTZ',
-  ],
-};
-
-export default AVAILABLE_CURRENCIES;
