+++ packages/core-pricing/plugins/product-price-coinbase-exchange.js
@@ -1,14 +1,39 @@
+import fetch from 'isomorphic-unfetch';
 import {
   ProductPricingDirector,
   ProductPricingAdapter,
 } from 'meteor/unchained:core-pricing';
 import { Countries } from 'meteor/unchained:core-countries';
+import Cache from '../utils/cache';
 
-import getFiatexchangeRateForCrypto from '../utils/getFiatexchangeRateForCrypto';
-import AVAILABLE_CURRENCIES from '../constants/available-exchanges';
+const CACHE_PERIOD = 60 * 60 * 0.1; // 10 minutes
+const SUPPORTED_CURRENCIES = [
+  'BTC',
+  'ETH',
+  'XRP',
+  'USDT',
+  'BCH',
+  'BSV',
+  'LTC',
+  'EOS',
+  'BNB',
+  'XTZ',
+];
+const cache = new Cache(CACHE_PERIOD);
 
-class ProductPriceEUR2CryptoExchange extends ProductPricingAdapter {
-  static key = 'shop.unchained.pricing.price-EUR-crypto-exchange';
+const getFiatexchangeRateForCrypto = async (base, target) => {
+  const { data } = await cache.get(`${base}-${target}`, () =>
+    fetch(`https://api.coinbase.com/v2/exchange-rates?currency=${base}`, {
+      method: 'GET',
+    }).then((res) => res.json())
+  );
+  return data?.rates?.[target];
+};
+
+export default getFiatexchangeRateForCrypto;
+
+class ProductPriceCoinbaseExchange extends ProductPricingAdapter {
+  static key = 'shop.unchained.pricing.price-coinbase-exchange';
 
   static version = '1.0';
 
@@ -17,7 +42,7 @@ class ProductPriceEUR2CryptoExchange extends ProductPricingAdapter {
   static orderIndex = 1;
 
   static isActivatedFor({ currency }) {
-    return AVAILABLE_CURRENCIES.crypto.indexOf(currency.toUpperCase()) !== -1;
+    return SUPPORTED_CURRENCIES.indexOf(currency.toUpperCase()) !== -1;
   }
 
   async calculate() {
@@ -43,4 +68,4 @@ class ProductPriceEUR2CryptoExchange extends ProductPricingAdapter {
   }
 }
 
-ProductPricingDirector.registerAdapter(ProductPriceEUR2CryptoExchange);
+ProductPricingDirector.registerAdapter(ProductPriceCoinbaseExchange);
